Prepare Xerces adapter refactoring for #642
* move classes linked to Xerces to the o.e.l.extensions.xerces
* move XMLPositionUtility#toLSPPosition to AbstractLSPErrorReporter
