+++ org.eclipse.lemminx/src/main/java/org/eclipse/lemminx/services/extensions/xerces/LSPMessageFormatter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.eclipse.lemminx.services.extensions.diagnostics;
+package org.eclipse.lemminx.services.extensions.xerces;
 
 import static org.eclipse.lemminx.dom.parser.Constants.*;
 import static org.eclipse.lemminx.utils.StringUtils.getString;
